+++ packages/contracts/dollar/package.json
@@ -18,6 +18,7 @@
     "clean": "hardhat clean",
     "build": "TS_NODE_TRANSPILE_ONLY=1 npx hardhat compile",
     "purge": "yarn clean && rimraf ./node_modules",
+    "export": "hardhat export --export-all ./deployments.json",
     "test": "hardhat test",
     "test:coverage": "hardhat coverage",
     "lint": "run-s lint:* build",
