+++ packages/dapp/components/config/abis/UniswapV3Pool.json
