+++ packages/dapp/components/utils/contracts.ts
@@ -1,21 +1,32 @@
 import { ethers } from "ethers";
-import { simpleRpcProvider } from "utils/provider";
-import erc20ABI from "config/abi/erc20.json";
+import UniswapV2PairABI from "../config/abis/UniswapV2Pair.json"
+import UniswapV3PoolABI from "../config/abis/UniswapV3Pool.json"
+import ChainlinkPriceFeedABI from "../config/abis/ChainlinkPriceFeed.json"
+import ERC20ABI from "../config/abis/ERC20.json"
 
 const getContract = (
-    chainId: number,
     abi: any,
     address: string,
-    signer?: ethers.Signer | ethers.providers.Provider
+    provider: ethers.providers.Provider
 ) => {
-    const signerOrProvider = signer ?? simpleRpcProvider(chainId);
-    return new ethers.Contract(address, abi, signerOrProvider);
+    return new ethers.Contract(address, abi, provider);
 };
 
+export const getUniswapV2FactoryContract = (address: string, provider: ethers.providers.Provider) => {
+    return getContract(UniswapV2PairABI, address, provider);
+}
+
+export const getUniswapV3PoolContract = (address: string, provider: ethers.providers.Provider) => {
+    return getContract(UniswapV3PoolABI, address, provider);
+}
+
+export const getChainlinkPriceFeedContract = (address: string, provider: ethers.providers.Provider): ethers.Contract => {
+    return getContract(ChainlinkPriceFeedABI, address, provider);
+}
+
 export const getERC20Contract = (
-    chainId: number,
     address: string,
-    signer?: ethers.Signer | ethers.providers.Provider
+    provider: ethers.providers.Provider
 ) => {
-    return getContract(chainId, erc20ABI, address, signer);
+    return getContract(ERC20ABI, address, provider);
 };
